% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{evaljs}
\alias{evaljs}
\title{Send javascript expression to the browser}
\usage{
evaljs(...)
}
\arguments{
\item{...}{character: JS expression (will be pasted together with no whitespace)}
}
\description{
Send javascript expression to the browser
}
