% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vapourlayer.R
\name{vl_map_ui}
\alias{vl_map_ui}
\alias{vl_map_ui_postamble}
\alias{vl_map_server}
\title{Vapourlayer map shiny module}
\usage{
vl_map_ui(id, view_wh = c(40, 40))

vl_map_ui_postamble()

vl_map_server(
  id,
  image_wh = 3200,
  initial_view = list(tiles_per_side = 1L, extent = c(-1, 1, -1, 1) * 2e+07, res = 32000),
  layerdef,
  target_crs = "EPSG:3031",
  cache = TRUE
)
}
\arguments{
\item{id}{string: HTML element id}

\item{view_wh}{numeric: two-element vector giving the width and height of the map widget as percentages of the browser window width and height}

\item{image_wh}{integer: image size in pixels. Depending on the rendering mechanism, the native image size might not actually be this, but the browser will scale to this size}

\item{initial_view}{list: a named list with components:
* tiles_per_side integer: e.g. a value of 2 means that we will generate the image as 2x2 set of tiles
* extent numeric: the initial image extent c(xmin, xmax, ymin, ymax) in projected coordinates
* res numeric: the resolution (in m) to use for the image when shown at its intial extent}

\item{layerdef}{reactive: TBD}

\item{target_crs}{string: target projection CRS string}

\item{cache}{logical or cachem: either TRUE/FALSE to use/not the default memory-based cache, or a `cachem` object (as returned by e.g. [cachem::cache_mem()])}
}
\value{
The UI and server components of a Shiny module
}
\description{
Vapourlayer map shiny module
}
